pyproject.toml
[project]
name = "audio-fdtd"
version = "0.0.0"
requires-python = ">= 3.12"
dependencies = [
    "scipy",
    "matplotlib",
    "numpy",
    "pillow"
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
markers = [
    "visual: marks tests that display plots for visual inspection (deselect with '-k not visual')"
]